🏗️ Refactor: comment 상태 관리 및 API 커스텀 훅 폴더 구조 재구성
